# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LIFE WITH AI - EdTech Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all published courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseListResponse"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /courses/{id}/lessons:
    post:
      operationId: addLesson
      tags: [lessons]
      summary: Add a lesson to a course (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /courses/{id}/lessons/{lessonId}:
    patch:
      operationId: updateLesson
      tags: [lessons]
      summary: Update a lesson (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete a lesson (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /courses/{id}/progress:
    get:
      operationId: getCourseProgress
      tags: [courses]
      summary: Get user progress for a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressResponse"
    post:
      operationId: updateCourseProgress
      tags: [courses]
      summary: Update lesson completion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressResponse"
  /courses/{id}/stream/{lessonId}:
    get:
      operationId: getLessonStream
      tags: [courses]
      summary: Get secure streaming URL for a lesson video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── PDF STORE ────────────────────────────────────────────────────────────────
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── ADMIN ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Course:
      type: object
      required: [id, title, thumbnail, description, price, category, difficulty, instructor, isPublished, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        thumbnail:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        category:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        instructor:
          type: string
        totalLessons:
          type: integer
        totalDuration:
          type: integer
        isPublished:
          type: boolean
        rating:
          type: ["number", "null"]
        totalStudents:
          type: integer
        createdAt:
          type: string
          format: date-time
    CourseDetail:
      type: object
      required: [id, title, thumbnail, description, price, category, difficulty, instructor, lessons, isPublished, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        thumbnail:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        category:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        instructor:
          type: string
        totalDuration:
          type: integer
        isPublished:
          type: boolean
        rating:
          type: ["number", "null"]
        totalStudents:
          type: integer
        isPurchased:
          type: boolean
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
        createdAt:
          type: string
          format: date-time
    CourseInput:
      type: object
      required: [title, description, price, category, difficulty, instructor]
      properties:
        title:
          type: string
        thumbnail:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        category:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        instructor:
          type: string
        isPublished:
          type: boolean
    CourseUpdate:
      type: object
      properties:
        title:
          type: string
        thumbnail:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        category:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        instructor:
          type: string
        isPublished:
          type: boolean
    CourseListResponse:
      type: object
      required: [courses, total, page, pages]
      properties:
        courses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
    Lesson:
      type: object
      required: [id, title, duration, order]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        videoPublicId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        duration:
          type: integer
        order:
          type: integer
        isPreview:
          type: boolean
    LessonInput:
      type: object
      required: [title, duration, order]
      properties:
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        videoPublicId:
          type: string
        notes:
          type: string
        duration:
          type: integer
        order:
          type: integer
        isPreview:
          type: boolean
    LessonUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        videoPublicId:
          type: string
        notes:
          type: string
        duration:
          type: integer
        order:
          type: integer
        isPreview:
          type: boolean
    ProgressResponse:
      type: object
      required: [courseId, completedLessons, totalLessons, percentage]
      properties:
        courseId:
          type: string
        completedLessons:
          type: array
          items:
            type: string
        totalLessons:
          type: integer
        percentage:
          type: number
    ProgressInput:
      type: object
      required: [lessonId, completed]
      properties:
        lessonId:
          type: string
        completed:
          type: boolean
    StreamResponse:
      type: object
      required: [streamUrl]
      properties:
        streamUrl:
          type: string

    # ── PDFs ──────────────────────────────────────────────────────────────────
...
    Order:
      type: object
      required: [id, itemId, itemType, itemTitle, amount, status, createdAt]
      properties:
        id:
          type: string
        itemId:
          type: string
        itemType:
          type: string
          enum: [course, pdf]
        itemTitle:
          type: string
        itemThumbnail:
          type: ["string", "null"]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed]
        razorpayOrderId:
          type: ["string", "null"]
        razorpayPaymentId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ── ADMIN ─────────────────────────────────────────────────────────────────
...